+++ test/client.test.js
@@ -445,7 +445,7 @@ describe('Client', () => {
     })
 
     test('disable checking', async () => {
-      new Client({ ...util.getCfg(), notifyAboutNewVersion: false })
+      new Client({ ...util.getCfg(), checkNewVersion: false })
       await new Promise(resolve => {
         setTimeout(() => {
           expect(console.info).not.toBeCalled()