+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Improve types (errors.pargeJSON, errors.TooManyRequests, `signal` for client.query options)
 - Fix false positive cases for ClientClosed error
-- Option to disable checking driver version (new Client({notifyAboutNewVersion: false}))
+- Option to disable checking driver version (new Client({checkNewVersion: false}))
 - Fix PageHelper honors size with map/filter
 - Fix module issue for `Vitessedge`